+++ ext/Server/Bot.lua
@@ -88,6 +88,8 @@ function Bot:__init(p_Player)
 	self._FullVehicleSteering = false
 	self._VehicleDirBackPositive = false
 	self._JetAbortAttackActive = false
+	self._LastVehicleHealth = 0.0
+	self._RepairVehicleEntity = nil
 	-- PID Controllers
 	-- normal driving
 	self._Pid_Drv_Yaw = PidController(5, 0.05, 0.2, 1.0)
@@ -166,6 +168,7 @@ function Bot:OnUpdatePassPostFrame(p_DeltaTime)
 							self:_UpdateAttacking()
 							if self._ActiveAction == BotActionFlags.ReviveActive or
 							self._ActiveAction == BotActionFlags.EnterVehicleActive or
+							self._ActiveAction == BotActionFlags.RepairActive or
 							self._ActiveAction == BotActionFlags.C4Active then
 								self:_UpdateMovementSprintToTarget()
 							else
@@ -287,6 +290,16 @@ function Bot:Revive(p_Player)
 	end
 end
 
+function Bot:Repair(p_Player)
+	if self.m_Kit == BotKits.Engineer and p_Player.soldier ~= nil and p_Player.controlledControllable ~= nil then
+		self._ActiveAction = BotActionFlags.RepairActive
+		self._RepairVehicleEntity = p_Player.controlledControllable
+		self._LastVehicleHealth = 0.0
+		self._ShootPlayer = nil
+		self._ShootPlayerName = p_Player.name
+	end
+end
+
 function Bot:ResetVehicleTimer()
 	self._VehicleWaitTimer = 0
 end
@@ -298,10 +311,45 @@ function Bot:EnterVehicleOfPlayer(p_Player)
 	self._ShootModeTimer = 0
 end
 
+function Bot:UpdateObjective(p_Objective)
+	local s_AllObjectives = m_NodeCollection:GetKnownObjectives()
+	for l_Objective,_ in pairs(s_AllObjectives) do
+		if l_Objective == p_Objective then
+			self:SetObjective(p_Objective)
+			break
+		end
+	end
+end
+
+function Bot:DeployIfPossible()
+	-- deploy from time to time
+	if self.m_Kit == BotKits.Support or self.m_Kit == BotKits.Assault then
+		if self.m_PrimaryGadget.type == WeaponTypes.Ammobag or self.m_PrimaryGadget.type == WeaponTypes.Medkit then
+			self:_AbortAttack()
+			self._WeaponToUse = BotWeapons.Gadget1
+			self._DeployTimer = 0
+		end
+	end
+end
+
+function Bot:ExitVehicle()
+	self.m_Player:ExitVehicle(false, false)
+	local s_Node = g_GameDirector:FindClosestPath(self.m_Player.soldier.worldTransform.trans, false)
+
+	if s_Node ~= nil then
+		-- switch to vehicle
+		self._InvertPathDirection = false
+		self._PathIndex = s_Node.PathIndex
+		self._CurrentWayPoint = s_Node.PointIndex
+		self._LastWayDistance = 1000
+	end
+end
+
 function Bot:IsReadyToAttack()
 	if
 	self._ActiveAction == BotActionFlags.OtherActionActive or
 	self._ActiveAction == BotActionFlags.ReviveActive or
+	self._ActiveAction == BotActionFlags.RepairActive or
 	self._ActiveAction == BotActionFlags.EnterVehicleActive or
 	self._ActiveAction == BotActionFlags.GrenadeActive then
 		return false
@@ -582,7 +630,14 @@ function Bot:SetSpeed(p_Speed)
 end
 
 function Bot:SetObjective(p_Objective)
-	self._Objective = p_Objective or ''
+	if self._Objective ~= p_Objective then
+		self._Objective = p_Objective or ''
+		local s_Point = m_NodeCollection:Get(self._CurrentWayPoint, self._PathIndex)
+		if s_Point ~= nil then
+			local s_Direction = m_NodeCollection:ObjectiveDirection(s_Point, self._Objective, self.m_InVehicle)
+			self._InvertPathDirection = (s_Direction == 'Previous')
+		end
+	end
 end
 
 function Bot:GetObjective()
@@ -1002,7 +1057,7 @@ function Bot:_UpdateAiming()
 		return
 	end
 
-	if self._ActiveAction ~= BotActionFlags.ReviveActive then
+	if self._ActiveAction ~= BotActionFlags.ReviveActive and self._ActiveAction ~= BotActionFlags.RepairActive then
 		if not self._Shoot or self._ShootPlayer.soldier == nil or self.m_ActiveWeapon == nil then
 			return
 		end
@@ -1147,6 +1202,27 @@ function Bot:_UpdateAiming()
 		local s_WorseningValue = (math.random()*self._Skill/self._DistanceToPlayer) -- value scaled in offset in 1m
 		s_Yaw = s_Yaw + s_WorseningValue
 		s_Pitch = s_Pitch + s_WorseningValue
+
+		self._TargetPitch = s_Pitch
+		self._TargetYaw = s_Yaw
+
+	elseif self._ActiveAction == BotActionFlags.RepairActive then -- repair
+		if self._ShootPlayer == nil or self._ShootPlayer.soldier == nil or self._RepairVehicleEntity == nil then
+			return
+		end
+		local s_PositionTarget = self._RepairVehicleEntity.transform.trans:Clone() -- aim at vehicle
+		local s_PositionBot = self.m_Player.soldier.worldTransform.trans:Clone() + m_Utilities:getCameraPos(self.m_Player, false, false)
+
+		local s_DifferenceZ = s_PositionTarget.z - s_PositionBot.z
+		local s_DifferenceX = s_PositionTarget.x - s_PositionBot.x
+		local s_DifferenceY = s_PositionTarget.y - s_PositionBot.y
+
+		local s_AtanDzDx = math.atan(s_DifferenceZ, s_DifferenceX)
+		local s_Yaw = (s_AtanDzDx > math.pi / 2) and (s_AtanDzDx - math.pi / 2) or (s_AtanDzDx + 3 * math.pi / 2)
+
+		--calculate pitch
+		local s_Distance = math.sqrt(s_DifferenceZ ^ 2 + s_DifferenceX ^ 2)
+		local s_Pitch = math.atan(s_DifferenceY, s_Distance)
 
 		self._TargetPitch = s_Pitch
 		self._TargetYaw = s_Yaw
@@ -1630,7 +1706,7 @@ function Bot:_UpdateWeaponSelection()
 					self.m_ActiveWeapon = self.m_SecondaryGadget
 					self._ShotTimer = - self:GetFirstShotDelay(self._DistanceToPlayer)
 				end
-			elseif (self._WeaponToUse == BotWeapons.Gadget1 and Config.BotWeapon == BotWeapons.Auto) or Config.BotWeapon == BotWeapons.Gadget1 then
+			elseif self._ActiveAction == BotActionFlags.RepairActive or (self._WeaponToUse == BotWeapons.Gadget1 and Config.BotWeapon == BotWeapons.Auto) or Config.BotWeapon == BotWeapons.Gadget1 then
 				if self.m_Player.soldier.weaponsComponent.currentWeaponSlot ~= WeaponSlot.WeaponSlot_2 and self.m_Player.soldier.weaponsComponent.currentWeaponSlot ~= WeaponSlot.WeaponSlot_4 then
 					self:_SetInput(EntryInputActionEnum.EIASelectWeapon4, 1)
 					self:_SetInput(EntryInputActionEnum.EIASelectWeapon3, 1)
@@ -1693,6 +1769,7 @@ function Bot:_UpdateDeployAndReload()
 	self._WeaponToUse = BotWeapons.Primary
 	self:_ResetActionFlag(BotActionFlags.C4Active)
 	self:_ResetActionFlag(BotActionFlags.ReviveActive)
+	self:_ResetActionFlag(BotActionFlags.RepairActive)
 	self:_ResetActionFlag(BotActionFlags.EnterVehicleActive)
 	self:_ResetActionFlag(BotActionFlags.GrenadeActive)
 	self:_AbortAttack()
@@ -1726,7 +1803,10 @@ function Bot:_UpdateDeployAndReload()
 end
 
 function Bot:_UpdateAttacking()
-	if self._ShootPlayer.soldier ~= nil and self._ActiveAction ~= BotActionFlags.EnterVehicleActive and self._Shoot then
+	if self._ShootPlayer.soldier ~= nil and
+	self._ActiveAction ~= BotActionFlags.EnterVehicleActive and
+	self._ActiveAction ~= BotActionFlags.RepairActive and 
+	self._Shoot then
 		if (self._ShootModeTimer < Config.BotFireModeDuration) or 
 			(Config.ZombieMode and self._ShootModeTimer < (Config.BotFireModeDuration * 4)) then
 
@@ -1957,13 +2037,14 @@ function Bot:_UpdateAttacking()
 			self:_AbortAttack()
 			self:_ResetActionFlag(BotActionFlags.ReviveActive)
 		end
-	elseif self._ActiveAction == BotActionFlags.EnterVehicleActive and self._ShootPlayer.soldier ~= nil then
+	-- enter vehicle
+	elseif self._ActiveAction == BotActionFlags.EnterVehicleActive and self._ShootPlayer ~= nil and self._ShootPlayer.soldier ~= nil then
 		self._ShootModeTimer = self._ShootModeTimer + Registry.BOT.BOT_UPDATE_CYCLE
 		self.m_ActiveMoveMode = BotMoveModes.ReviveC4 -- movement-mode : revive
 
 		--check for enter of vehicle if close
 		if self._ShootPlayer.soldier.worldTransform.trans:Distance(self.m_Player.soldier.worldTransform.trans) < 5 then
-			self:_EnterVehicle()
+			self:_EnterVehicle(true)
 			self._TargetPitch = 0.0
 			self:_AbortAttack()
 			self:_ResetActionFlag(BotActionFlags.EnterVehicleActive)
@@ -1973,6 +2054,32 @@ function Bot:_UpdateAttacking()
 			self:_AbortAttack()
 			self:_ResetActionFlag(BotActionFlags.EnterVehicleActive)
 		end
+	-- repair
+	elseif self._ActiveAction == BotActionFlags.RepairActive and self._ShootPlayer ~= nil and self._ShootPlayer.soldier ~= nil and self._RepairVehicleEntity ~= nil then
+		self._ShootModeTimer = self._ShootModeTimer + Registry.BOT.BOT_UPDATE_CYCLE
+		self.m_ActiveMoveMode = BotMoveModes.ReviveC4 -- movement-mode : repair
+
+		local s_CurrentHealth = PhysicsEntity(self._RepairVehicleEntity).internalHealth
+
+		--check for repair if close to vehicle
+		if self._RepairVehicleEntity.transform.trans:Distance(self.m_Player.soldier.worldTransform.trans) < 5 then
+			if s_CurrentHealth ~= self._LastVehicleHealth then
+				self._ShootModeTimer = Registry.BOT.MAX_TIME_TRY_REPAIR - 2.0  -- continue for few seconds on progress
+			end
+			self._LastVehicleHealth = s_CurrentHealth
+			self._TargetPitch = 0.0
+			self._AttackModeMoveTimer = 0 -- don't jump anymore
+			self:_SetInput(EntryInputActionEnum.EIAFire, 1)
+		end
+
+		-- Abort conditions
+		if self._ShootModeTimer > Registry.BOT.MAX_TIME_TRY_REPAIR or self._RepairVehicleEntity == nil then -- abort this after some time
+			self._TargetPitch = 0.0
+			self:_AbortAttack()
+			self:_ResetActionFlag(BotActionFlags.RepairActive)
+			self._WeaponToUse = BotWeapons.Primary
+		end
+	
 	elseif self._ShootPlayer.soldier == nil then -- reset if enemy is dead
 		self:_AbortAttack()
 	end
@@ -2046,13 +2153,13 @@ function Bot:_UpdateVehicleMovableId()
 	end
 end
 
-function Bot:_EnterVehicleEntity(p_Entity)
+function Bot:_EnterVehicleEntity(p_Entity, p_PlayerIsDriver)
 	if p_Entity ~= nil then
 		local s_Position = p_Entity.transform.trans
 
 		-- keep one seat free, if enough available
 		local s_MaxEntries = p_Entity.entryCount
-		if s_MaxEntries > 2 then
+		if not p_PlayerIsDriver and s_MaxEntries > 2 then
 			s_MaxEntries = s_MaxEntries -1
 		end
 	
@@ -2093,7 +2200,7 @@ function Bot:_EnterVehicleEntity(p_Entity)
 	end
 end
 
-function Bot:_EnterVehicle(p_Name)
+function Bot:_EnterVehicle(p_PlayerIsDriver)
 	local s_Iterator = EntityManager:GetIterator("ServerVehicleEntity")
 	local s_Entity = s_Iterator:Next()
 
@@ -2110,7 +2217,7 @@ function Bot:_EnterVehicle(p_Name)
 		s_Entity = s_Iterator:Next()
 	end
 	if s_ClosestEntity ~= nil then
-		return self:_EnterVehicleEntity(s_ClosestEntity)
+		return self:_EnterVehicleEntity(s_ClosestEntity, p_PlayerIsDriver)
 	end
 	return -3 -- no vehicle found
 end
@@ -2165,19 +2272,6 @@ function Bot:_UpdateStaticMovement()
 
 		self._TargetYaw = self._TargetPlayer.input.authoritativeAimingYaw + ((self._TargetPlayer.input.authoritativeAimingYaw > math.pi) and -math.pi or math.pi)
 		self._TargetPitch = self._TargetPlayer.input.authoritativeAimingPitch
-	end
-end
-
-function Bot:ExitVehicle()
-	self.m_Player:ExitVehicle(false, false)
-	local s_Node = g_GameDirector:FindClosestPath(self.m_Player.soldier.worldTransform.trans, false)
-
-	if s_Node ~= nil then
-		-- switch to vehicle
-		self._InvertPathDirection = false
-		self._PathIndex = s_Node.PathIndex
-		self._CurrentWayPoint = s_Node.PointIndex
-		self._LastWayDistance = 1000
 	end
 end
 
@@ -2506,7 +2600,7 @@ function Bot:_UpdateNormalMovement()
 			if s_Point.Data ~= nil and s_Point.Data.Action ~= nil then
 				if s_Point.Data.Action.type == "vehicle" then
 					if Config.UseVehicles then
-						local s_RetCode, s_Position = self:_EnterVehicle()
+						local s_RetCode, s_Position = self:_EnterVehicle(false)
 						if s_RetCode == 0 then
 							self:_ResetActionFlag(BotActionFlags.OtherActionActive)
 							local s_Node = g_GameDirector:FindClosestPath(s_Position, true)
@@ -2589,6 +2683,12 @@ function Bot:_UpdateNormalMovement()
 				end
 				self:_SetInput(EntryInputActionEnum.EIAStrafe, self.m_StrafeValue)
 				self._SidewardsTimer = self._SidewardsTimer - Registry.BOT.BOT_UPDATE_CYCLE
+			end
+
+			-- use parachute if needed
+			local s_VelocityFalling = PhysicsEntity(self.m_Player.soldier).velocity.y
+			if s_VelocityFalling < -50.0 then
+				self:_SetInput(EntryInputActionEnum.EIAToggleParachute, 1)
 			end
 
 			local s_DifferenceY = s_Point.Position.z - self.m_Player.soldier.worldTransform.trans.z